requirements.txt
# llama-index 
llama-index==0.8.48

# llama-hub
llama-hub==0.0.26

# database clients
weaviate-client==3.23.2
pymongo==4.5.0

# entity extractor
span-marker==1.3.0
pypdf==3.15.4
docx2txt==0.8.0
protobuf==4.24.2
PyMuPDF==1.23.4
pytesseract==0.3.10

# nlp related, embeddings etc.
sentence-transformers==2.2.2
bitsandbytes==0.41.1


# api and data models
fastapi==0.103.2
pydantic-settings==2.0.3

# frontend
streamlit==1.27.2


## llama-cpp-python
# CMAKE_ARGS="-DLLAMA_CUBLAS=on" FORCE_CMAKE=1 pip install llama-cpp-python==0.2.11 --force-reinstall --upgrade 